Move favourites drawer to component and Context API
